+++ src/Collection/CHANGELOG.md
@@ -22,6 +22,18 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Nothing
 -->
 
+## [1.2.0] - 2019-11-08
+### Added
+- Huge update for `TypedCollection`:
+    - check type on item add to collection
+    - add `downgrade` method that returns simple `Collection` instance
+    - override some methods witch logic requires downgrade, restore original `TypedCollection` when we can
+    - cover `BaseCollection` and `TypedCollection` with tests
+
+## [1.1.1] - 2019-10-18
+### Fixed
+- Fix checking class and interface existence in TypedCollection
+
 ## [1.1.0] - 2019-10-18
 ### Added
 - Add TypedCollection